requirements.txt
functions-framework==3.*
google-api-python-client==2.94.0
google-cloud-storage==2.11.0
Pillow==9.5.0
reportlab==3.6.13